+++ db/migrate/20220215002456_fix_on_delete_versions.rb
@@ -1,0 +1,6 @@
+class FixOnDeleteVersions < ActiveRecord::Migration[6.1]
+  def change
+    remove_foreign_key :versions, :projects
+    add_foreign_key :versions, :projects, on_delete: :cascade
+  end
+end